Sql Builds, queries and ERD
SQL to build the tables, the Queries, the ERD generated from within the sql progress tool.  I can;t get data to load to the two tables in which the the two columns are foreign keys.  I created a combined key to make them unique, but not only can I not load data, I can not retrieve data from those tables.